pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "Long-running local AI assistant driven by Feishu, LangGraph, and local skills."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "langgraph>=1.1.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic-settings>=2.7.0",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
jarvis = "app.main:main"

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
